requirement.txt
asgiref==3.5.0
certifi==2021.10.8
charset-normalizer==2.0.12
Django==4.0.2
django-parler==2.3
django-rosetta==0.9.8
idna==3.3
polib==1.1.1
psycopg2-binary==2.9.3
requests==2.27.1
sqlparse==0.4.2
urllib3==1.26.8
